Error Page Added.
Error page using a modified version of "canvas.js"

Also fixed the beard logo which was not being served properly because it
was svgz. changing to svg fixed it. (I hope.)

Icon was not showing up in title bar, never noticed before because it
was cached before I changed the image location.