Fix Most IE issues.
Working in chrome originally everything was as expected.

Tested in IE11 to find no styles were being loaded.
apparently there's a difference between "css/text" and "text/css" Though
no google search would tell me that.

Everything but the svg backgrounds and loaded fonts are working now in
IE11, I think the background might be fixed when it's uploaded though,
so this is also a test for that.

I do not know why the loaded fonts don't work, I thought TTF was
supported in IE11. Whatever, impact as a backup looks okay.