Tiny changes.
This is two seconds after the previous push. I wish I had done these
changes five seconds earlier.

"Submit Query" button now says "Send Message"

Also added the link that explained the font-face rule to me because it
was well written and It may come in handy in the future.